// src/enums/category.enums.ts
export enum CategoryText {
  SECTION_TITLE = "Our Services",
  SECTION_DESCRIPTION = "End-to-end logistics made simple: from warehousing to reverse logistics - all handled with precision and speed.",
  EXPLORE_BUTTON = "Explore",
  EST_PREFIX = "EST. 2020 /",
}

export enum CategoryTitle {
  WAREHOUSING = "Warehousing & Fulfillment",
  ECOMMERCE_FULFILLMENT = "E-commerce Fulfillment",
  SHIPPING_FREIGHT_MANAGEMENT = "Shipping & Freight Management",
  REVERSE_LOGISTICS = "Reverse Logistics",
}