// src/enums/navigation.enums.ts
export enum NavItems {
    OUR_SERVICES = "Our Services",
    HOW_WE_WORK = "How We Work",
    WHY_CHOOSE_US = "Why Choose Us",
    ABOUT_US = "About Us",
    FAQS = "FAQs",
    CONTACT_US = "Contact Us",
    PRIVACY_POLICY = "Privacy Policy",
    TERMS_OF_SERVICE = "Terms of Service",
    COOKIE_POLICY = "Cookie Policy",
}